AtualizacaoBancoDeDados

+++ src/database/mysql/migrations/20231018165113_accountUser.js
@@ -5,12 +5,12 @@
 exports.up = async function(knex) {
     await knex.schema.createTable("accountUser", table => {
         table.increments("id")
-        table.integer("amount")
-        table.integer("invoice_amount")
+        table.float("amount")
+        table.float("invoice_amount")
         table.integer("payment_data")
         table.integer("invoice_closing")
-        table.integer("credit_card")
-        table.integer("loan")
+        table.float("credit_card")
+        table.float("loan")
         table.integer("user_id").references("id").inTable("users").onDelete("CASCADE")
     })
 };
